requirements.txt
PyQt5>=5.15.0
pyserial>=3.5
pandas>=1.3.0
pyqtgraph>=0.12.0
numpy>=1.21.0
bitstring>=3.1.0
